Users, Posts sections are 100% completed and Category addition feature added in the application

+++ codehacking/app/Http/Controllers/AdminUsersController.php
@@ -3,13 +3,14 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use App\Http\Requests\PostsCreateRequest;
+use App\Http\Requests;
 use App\User;
 use App\Photo;
 use App\Role;
 use App\Post;
-use App\Http\Requests;
-use App\Http\Requests\UsersRequest;
-use App\Http\Requests\UsersEditRequest;
+use App\Category;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Session;
 
 class AdminUsersController extends Controller
@@ -101,7 +102,10 @@ class AdminUsersController extends Controller
     public function destroy($id)
     {
         $user = User::findOrFail($id);
-        unlink(public_path() . $user->photo['file']);
+        if($user->photo['file'])
+        {
+          unlink(public_path() . $user->photo['file']);
+        }
         $user->delete();
         Session::flash('deleted_user', 'The user has been deleted');
         return redirect('/admin/users');
